requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.12" and python_version < "4.0"
boto3==1.34.157 ; python_version >= "3.12" and python_version < "4.0"
botocore==1.34.157 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.6.2 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
dill==0.3.8 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
ecs-logging==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
elastic-transport==8.13.1 ; python_version >= "3.12" and python_version < "4.0"
elasticsearch==8.15.0 ; python_version >= "3.12" and python_version < "4.0"
frozendict==2.4.4 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
logfmter==0.0.6 ; python_version >= "3.12" and python_version < "4.0"
lxml==5.2.2 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
multitasking==0.0.11 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
openai==1.40.2 ; python_version >= "3.12" and python_version < "4.0"
pandarallel==1.6.5 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
peewee==3.17.5 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.2.2 ; python_version >= "3.12" and python_version < "4.0"
polygon-api-client==1.13.7 ; python_version >= "3.12" and python_version < "4.0"
psutil==6.0.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.18.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.7.4 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
retrying==1.3.4 ; python_version >= "3.12" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.12" and python_version < "4.0"
s3transfer==0.10.2 ; python_version >= "3.12" and python_version < "4.0"
setuptools==72.1.0 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.12" and python_version < "4.0"
typer==0.12.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==1.26.19 ; python_version >= "3.12" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.12" and python_version < "4.0"
yfinance==0.2.40 ; python_version >= "3.12" and python_version < "4.0"
